feat: no-reply/no-match improvements (VF-1997)

+++ lib/services/runtime/handlers/interaction.ts
@@ -1,35 +1,40 @@
 import { Node as BaseNode, Trace } from '@voiceflow/base-types';
-import { IntentEvent } from '@voiceflow/base-types/build/node/utils';
 import { Node as ChatNode } from '@voiceflow/chat-types';
 import { Node as GeneralNode } from '@voiceflow/general-types';
 
 import { Action, HandlerFactory } from '@/runtime';
 
 import { StorageType } from '../types';
-import { addButtonsIfExists, addRepromptIfExists } from '../utils';
+import { addButtonsIfExists } from '../utils';
 import CommandHandler from './command';
 import { findEventMatcher } from './event';
 import NoMatchHandler from './noMatch';
+import NoReplyHandler from './noReply';
 import RepeatHandler from './repeat';
 
 const utilsObj = {
   repeatHandler: RepeatHandler(),
   commandHandler: CommandHandler(),
   noMatchHandler: NoMatchHandler(),
+  noReplyHandler: NoReplyHandler(),
   findEventMatcher,
   addButtonsIfExists,
-  addRepromptIfExists,
 };
 
 export const InteractionHandler: HandlerFactory<GeneralNode.Interaction.Node | ChatNode.Interaction.Node, typeof utilsObj> = (utils) => ({
   canHandle: (node) => !!node.interactions,
-  handle: (node, runtime, variables, program) => {
+  handle: (node, runtime, variables) => {
+    if (runtime.getRequest() === null) {
+      return utils.noReplyHandler.handle(node, runtime, variables);
+    }
+
     if (runtime.getAction() === Action.RUNNING) {
       utils.addRepromptIfExists(node, runtime, variables);
       utils.addButtonsIfExists(node, runtime, variables);
 
-      // clean up no matches counter on new interaction
+      // clean up no-matches and no-replies counters on new interaction
       runtime.storage.delete(StorageType.NO_MATCHES_COUNTER);
+      runtime.storage.delete(StorageType.NO_REPLIES_COUNTER);
 
       // quit cycleStack without ending session by stopping on itself
       return node.id;
@@ -44,10 +49,10 @@ export const InteractionHandler: HandlerFactory<GeneralNode.Interaction.Node | C
         // allow handler to apply side effects
         matcher.sideEffect();
 
-        if ((event as IntentEvent).goTo) {
+        if ((event as BaseNode.Utils.IntentEvent).goTo) {
           runtime.trace.addTrace<Trace.GoToTrace>({
             type: BaseNode.Utils.TraceType.GOTO,
-            payload: { request: (event as IntentEvent).goTo!.request },
+            payload: { request: (event as BaseNode.Utils.IntentEvent).goTo!.request },
           });
 
           // stop on itself to await for new intent request coming in
@@ -67,20 +72,12 @@ export const InteractionHandler: HandlerFactory<GeneralNode.Interaction.Node | C
     if (utils.commandHandler.canHandle(runtime)) {
       return utils.commandHandler.handle(runtime, variables);
     }
+
     if (utils.repeatHandler.canHandle(runtime)) {
       return utils.repeatHandler.handle(runtime);
     }
 
-    // check for noMatches to handle
-    if (utils.noMatchHandler.canHandle(node, runtime, variables, program)) {
-      return utils.noMatchHandler.handle(node, runtime, variables, program);
-    }
-
-    runtime.trace.addTrace<Trace.PathTrace>({
-      type: BaseNode.Utils.TraceType.PATH,
-      payload: { path: 'choice:else' },
-    });
-    return node.elseId || null;
+    return utils.noMatchHandler.handle(node, runtime, variables);
   },
 });
